+++ src/components/UpdateServiceForm.tsx
@@ -78,11 +78,11 @@ const UpdateServiceForm: React.FC<EditServiceModalProps> = ({ isOpen, onClose, s
         client_id: service.client_id,
         items: service.items,
         estimated_completion_date: service.estimated_completion_date,
-        is_weight: service.is_weight, // Assuming you have a way to determine this from service data
-        is_piece: service.is_piece, // Assuming you have a way to determine this from service data
-        is_paid: service.is_paid, // Assuming you have a way to determine this from service data
-        completed_at: service.completed_at, // Assuming you have a way to determine this from service data
-        weight: service.weight, // Assuming you have a way to determine this from service data
+        is_weight: service.is_weight,
+        is_piece: service.is_piece,
+        is_paid: service.is_paid,
+        completed_at: service.completed_at,
+        weight: service.weight,
     });
     const [clients, setClients] = useState<SelectOption[]>([]);
     const [selectedItemsWithOptions, setSelectedItemsWithOptions] = useState<Array<SelectedItemWithOption>>([]);
@@ -90,7 +90,18 @@ const UpdateServiceForm: React.FC<EditServiceModalProps> = ({ isOpen, onClose, s
 
     // Função para carregar clientes e converter para o formato esperado por react-select
     const fetchClients = async () => {
-        const response = await fetch('http://localhost:8080/clients');
+        const token = localStorage.getItem('token'); // Retrieve the token from localStorage
+
+        if (!token) {
+            console.error('No token found, please login first');
+            return;
+        }
+        const response = await fetch('http://localhost:8080/clients', {
+            headers: {
+                'Authorization': `Bearer ${token}`,
+                'Content-Type': 'application/json',
+            },
+        });
         const data = await response.json(); // Supondo que `Client` seja o tipo dos seus clientes
         const responseClients: Client[] = data.clients
         const clientOptions: SelectOption[] = responseClients.map(client => ({
@@ -133,10 +144,18 @@ const UpdateServiceForm: React.FC<EditServiceModalProps> = ({ isOpen, onClose, s
             completed_at: formData.completed_at
         };
 
+        const token = localStorage.getItem('token'); // Retrieve the token from localStorage
+
+        if (!token) {
+            console.error('No token found, please login first');
+            return;
+        }
+
         // Send PATCH request to update service details
         const serviceResponse = await fetch(`http://localhost:8080/services/${serviceId}`, {
             method: 'PUT',
             headers: {
+                'Authorization': `Bearer ${token}`,
                 'Content-Type': 'application/json',
             },
             body: JSON.stringify(servicePayload),
@@ -149,14 +168,14 @@ const UpdateServiceForm: React.FC<EditServiceModalProps> = ({ isOpen, onClose, s
 
             try {
                 const errorResponse = await serviceResponse.json();
-                if (errorResponse.error === "Validation failed") { 
-                  toast.error(errorResponse.details.message);
+                if (errorResponse.error === "Validation failed") {
+                    toast.error(errorResponse.details.message);
                 } else {
-                  toast.error("An error occurred. Please try again.");
+                    toast.error("An error occurred. Please try again.");
                 }
-              } catch (error) {
+            } catch (error) {
                 toast.error("An unexpected error occurred. Please try again.");
-              }
+            }
         } else {
             toast.success("Serviço atualizado com sucesso!")
         }
@@ -173,6 +192,7 @@ const UpdateServiceForm: React.FC<EditServiceModalProps> = ({ isOpen, onClose, s
             const itemsResponse = await fetch(`http://localhost:8080/services/${serviceId}/items`, {
                 method: 'POST',
                 headers: {
+                    'Authorization': `Bearer ${token}`,
                     'Content-Type': 'application/json',
                 },
                 body: JSON.stringify({ items: newItems }),
@@ -221,9 +241,17 @@ const UpdateServiceForm: React.FC<EditServiceModalProps> = ({ isOpen, onClose, s
             item_quantity: newQuantity
         };
 
+        const token = localStorage.getItem('token'); // Retrieve the token from localStorage
+
+        if (!token) {
+            console.error('No token found, please login first');
+            return;
+        }
+
         const response = await fetch(`http://localhost:8080/services/${serviceId}/items/${itemId}`, {
             method: 'PATCH',
             headers: {
+                'Authorization': `Bearer ${token}`,
                 'Content-Type': 'application/json',
             },
             body: JSON.stringify(payload),
@@ -240,8 +268,19 @@ const UpdateServiceForm: React.FC<EditServiceModalProps> = ({ isOpen, onClose, s
 
 
     const removeItem = async (indexToRemove: number, serviceId: string, serviceItemID: string) => {
+        const token = localStorage.getItem('token'); // Retrieve the token from localStorage
+
+        if (!token) {
+            console.error('No token found, please login first');
+            return;
+        }
+
         const response = await fetch(`http://localhost:8080/services/${serviceId}/items/${serviceItemID}`, {
             method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${token}`,
+                'Content-Type': 'application/json',
+            },
         });
 
         if (response.ok) {
@@ -301,8 +340,20 @@ const UpdateServiceForm: React.FC<EditServiceModalProps> = ({ isOpen, onClose, s
 
     useEffect(() => {
         const fetchService = async () => {
+            const token = localStorage.getItem('token'); // Retrieve the token from localStorage
+
+            if (!token) {
+                console.error('No token found, please login first');
+                return;
+            }
+
             try {
-                const response = await fetch(`http://localhost:8080/services/${serviceId}`);
+                const response = await fetch(`http://localhost:8080/services/${serviceId}`, {
+                    headers: {
+                        'Authorization': `Bearer ${token}`,
+                        'Content-Type': 'application/json',
+                    },
+                });
                 if (!response.ok) {
                     throw new Error('Network response was not ok');
                 }
